Opt into go modules
Down stream projects need to vendor jmespath dependencies because
jmespath hasn't opted into go modules yet.

This is my interpretation of:

If a dependency of your module does not itself have a go.mod (e.g.,
because the dependency has not yet opted in to modules itself), or if
its go.mod file is missing one or more of its dependencies (e.g.,
because the module author did not run go mod tidy), then the missing
transitive dependencies will be added to your module's requirements,
along with an // indirect comment to indicate that the dependency is not
from a direct import within your module.

From https://github.com/golang/go/wiki/Modules#why-does-go-mod-tidy-record-indirect-and-test-dependencies-in-my-gomod

And my experience using aws-sdk-go as it depends on go-jamespath.