Add github.com/stretchr/testify@v1.5.1 as an internal submodule
Adds the testify testing framework as an internal Go submodule. The Go
submodule import paths are all rewritten to refer to its self, and had
its gopkg.in/yaml.v2 dependency updated to a that is not vulnerable to
CVE-2019-11254.

The submodule was created instead of updating the testify dependency,
because testify@v1.6.0 which updated to yaml.v3 is not compatible with
Go versions < v1.12.